git-svn-id: svn://123.57.173.36/meijia-parent@2045 52ac4245-7803-4e87-b018-04b7c5ce731e

+++ simi-dao/src/main/java/com/simi/po/dao/xcloud/XcompanySettingMapper.java
@@ -8,9 +8,9 @@ import com.simi.vo.xcloud.CompanySettingSearchVo;
 public interface XcompanySettingMapper {
     int deleteByPrimaryKey(Long id);
 
-    int insert(XcompanySetting record);
+    Long insert(XcompanySetting record);
 
-    int insertSelective(XcompanySetting record);
+    Long insertSelective(XcompanySetting record);
 
     XcompanySetting selectByPrimaryKey(Long id);
